requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
lingua-language-detector>=2.0.0
pydantic>=2.0.0
